+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 * `CocoDataset.add_track` will now pick a name using a uuid if one is unspecified.
+* Added more metrics to outputs in coco eval
 
 ### Fixed
 * CIFAR grabdata now respects dpath